pyproject.toml
[project]
name = "gymshark-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-beam>=2.63.0",
    "google-cloud-bigquery>=3.30.0",
    "google-cloud-storage>=3.1.0",
    "mock>=5.1.0",
    "pandas>=2.2.3",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
]
